Implement [] Remove all the Implementations
