+++ Switcher/AppDelegates/AppDelegate.swift
@@ -5,7 +5,7 @@
 //  Created by 김수환 on 2023/06/10.
 //
 
-import AppKit
+import SwiftUI
 
 class AppDelegate: NSObject, NSApplicationDelegate {
     
@@ -19,7 +19,6 @@ class AppDelegate: NSObject, NSApplicationDelegate {
         }
         createEventTap()
         makeMenuButton()
-        setKeyDefaultValue()
     }
     
     @objc func MenuButtonToggle(sender: AnyObject) {
@@ -33,7 +32,20 @@ class AppDelegate: NSObject, NSApplicationDelegate {
     }
     
     func applicationShouldTerminateAfterLastWindowClosed(_ sender: NSApplication) -> Bool {
-        keyMapWindow != nil ? (keyMapWindow = nil) : nil
         return false
+    }
+}
+
+
+// MARK: - Open In Window
+
+private extension View {
+    
+    func openInWindow(title: String, sender: Any?) -> NSWindow {
+        let win = NSWindow(contentViewController: NSHostingController(rootView: self))
+        win.title = title
+        win.makeKeyAndOrderFront(sender)
+        win.orderFrontRegardless()
+        return win
     }
 }
